+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* Experimental scripts for loading and saving new records from and to a records store (essentially renaming their files) 
+
 ## [0.2.0] - 2024-11-04
 
 ### Removed [BREAKING!]