+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 * set transfer opts script setting all options with values from the last record
+* set transfer opts script setting opts that already exist
 
 ## [0.2.0] - 2024-11-04