+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * date update handling
+* including additional text in citations
 
 ## [0.2.0] - 2024-11-04